requirements.txt
# Core Dependencies for Dr. Robert Young Semantic Search System

# HTTP requests library for web scraping and API calls
requests

# HTML parsing library for content extraction from web pages
beautifulsoup4

# MySQL database connector for data storage
mysql-connector-python

# Pre-trained transformer models for semantic embeddings
# Used for converting text to vector representations
sentence-transformers

# FastAPI web framework for building the REST API
fastapi

# ASGI server for running FastAPI applications
uvicorn

# Numerical computing library for vector operations
numpy
